requirements.txt
beautifulsoup4==4.12.2
lxml==5.0.0
python-dateutil==2.8.2
python-pptx==0.6.23
selenium==4.16.0
soupsieve==2.5
